requirements.txt
maltego-trx
pymisp
python-dotenv